requirements.txt
absl-py==0.7.1
aiofiles==0.4.0
aiohttp==3.5.4
alpaca-trade-api==0.26
astor==0.7.1
astroid==2.2.5
async-timeout==3.0.1
asyncio-nats-client==0.8.2
attrs==19.1.0
audioread==2.1.6
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coinbase==2.1.0
cycler==0.10.0
feedparser==5.2.1
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
google-api-core==1.9.0
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-bigquery==1.9.0
google-cloud-core==0.29.1
google-cloud-pubsub==0.40.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpc-google-iam-v1==0.11.4
grpcio==1.19.0
h5py==2.9.0
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
multidict==4.5.2
numpy==1.16.2
pafy==0.5.4
pandas==0.24.2
pandas-datareader==0.7.0
pbr==5.1.3
PlexAPI==3.1.0
protobuf==3.7.1
pyacoustid==1.1.5
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.8.1
pylint==2.3.1
pyparsing==2.4.0
pyPodcastParser==2.0.0
python-chess==0.28.0
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
rope==0.14.0
rsa==4.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
soupsieve==1.8
TA-Lib==0.4.17
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.31.1
typed-ast==1.3.1
urllib3==1.24.1
websocket-client==0.56.0
websockets==7.0
Werkzeug==0.14.1
wrapt==1.11.1
yarl==1.3.0
youtube-dl==2019.3.18
